176641412 fix add remove nodes (#46)
* Refactor code that determines the default partition

Always set the default partition to the "configurator"
partition by default and don't use the "configurator" partition
when the default_partition config is set.

* remove observation of joined event

* change the way the slurm-configurator charm status is set

The slurm-configurator status was previously determined only by
looking at what related units are available. We now take into
consideration which applications are joined separately from
which applications are available when generating the unit
status.

* check for relations are joined
* check for component availablity
* display a slurm-configurator status for joined vs available

* ignore pep3 C901 (too complex)

C901 is violated by the function that checks the connected
charm components and generates the status. Remove this check
as we intend for our status checker to be this way and do
not want an error.

* ensure we create a cache object with a unique key

The most recent built charms from the current ci run weren't
previously being cached because an item with the non-unique
name already exist. This ensures the built charms from the
current ci run are cached and used in testing.

* revert configless changes, save for different pr

* remove complexity in _check_status()

* change slurm-configurator active status message to something meaningful

* use private networks for created nat space

* get joined vs available status right

* run black and isort

* point all charms at slurm-ops-manager@configless

* add flake8 ignore for I201, I100, E203, E501, W503 to accommodate black+isort

* ensure unique inventory for each partition

* fix typo

* midnight sprint

* checkpoint

checkpoint where everything is working correctly except for
slurm_component_restart()

* support adding and removing nodes without restarting slurm

These changes introduce add/remove capability for slurm nodes
in a way that doesn't disrupt jobs.

Many additional changes and drive bys along the way, the primary
change that matters is that we do not restart slurmd anymore
following re-configuration activities. Instead we configure
slurm to run in configless mode and run 'scontrol reconfigure' in
order to get nodes to pick up configuration changes.

* fix lint errors

* fix typo

* give the cluster an extra few seconds to settle

* try using constraints

* specify model (-m) in deploy commands

* update github action to use bundle to deploy slurm

* carry new job name through

* add slurmrestd to focal-beta-bundle.yaml

* add focal lxd beta bundle and bring slurmrestd up to speed

Fixes pivotal story 176788868

These changes ensure that slurmrestd charm works as expected by
bringing forward patterns used throughout the rest of the charm
stack.

* correctly determine if the relation is joined

* pass the snapstore-channel config into upgrade()

* create centos7 bundle

* slurmrestd_requires.py -> inerface_slurmrestd.py

* bump slurm-ops-manager in charm requirements.txt

use the latest release of slurm-ops-manager

* bump requirements.txt

* ensure slurmd gets the slurm_config not just the partition

Co-authored-by: jamesbeedy <jamesbeedy>