slurmd: set partition name on peer relation data
This fixes a crash on all slurmd units when removing the leader of the
application. The partition name was stored in the leader, so removing it
looses this information and the slurm.conf is then scrambled.

This commit stores the partition name in the peer relation, so if the
leader is removed, we don't loose this information.

Add Bats test to assure partition names don't have spaces in them, as
spaces would break the slurm configuration file.