add charmcraft.yaml file
With charmcraft 1.1.0, we need another YAML file in the repository,
charmcraft.yaml, that describes how to build and which OSes the charm
run. This patch adds the YAML file to build the charm on Ubuntu Focal
and run on Ubuntu Focal and CentOS7.

Charmcraft pack creates a horrible filename. We copy the file from the
project-dir to the root with a simpler filename, as to make it still
compatible with slurm-bundles.