refactor configuration of partition names
This patch improves the code that sets the partition name of a Juju
slurmd application. It fixes a race condition where the partition-name
is set in a bundle YAML file, but it is not propagated to the charm
state.
It also changes the default partition name to osd-{appname} instead of
juju-compute-randomstring.