add influxdb and grafana relations to slurmctld
Add relation to infuxdb, to use the SLURM InfluxDB accounting plugin.
This plugin allows job profiling. Two new configuration options added:
- acct-gather-frequency: the sampling interval for the acct_gather
  plugins.
- acct-gather-custom: the custom values for the acct_gather.conf file.

Add relation to grafana. This relation depends on the influxdb relation:
when relating to grafana, send the information about the influxdb (host,
port, user, pass, bucket) as a data-source for dashboards.

Test we can alter the JobAcctGatherFrequency and propagate the changes
to all nodes.