Use etcd for internal node accounting
This patch adds an etcd instance running on slurmctld/leader to keep the
internal node accounting. etcd now keeps the list of nodes that
slurmctld is aware of. Each slurmd unit then queries the etcd server to
check if it is accounted for, instead of using Juju's relation data.

This is to reduce the number of events fired when adding/removing units.
Previously, each add-unit operation was firing a relation_changed twice
in each slurmd unit when a new one is added.