Store munge key in etcd
This patch stores the munge key in etcd database at key `munge/key`. This
allows a third party client to query the database for the munge key, in
order to authenticate an external application with Slurm.

This also changes etcd to always use auth, so slurmd must use the
`slurmd` account with a password generated randomly in slurmctld.
Slurmctld-charm sends it via the relation data.

The following new actions were created:
- slurmctld: etcd-create-munge-account to create an account able to
  query the munge key.
- slurmctld: etcd-get-root-password to get the password for the etcd
  `root` account
- slurmctld: etcd-get-slurmd-password to get the password for the etcd
  `slurmd` account