slurmctld: add support for TLS certificates
Add three configurations in slurmctld charm to allow the operator to
supply TLS certificates for etcd. Two options for the TLS cert (pub and
private files), and one optional as the CA public certificate, in case
the nodes do not have that or it was used an internal CA. etcd will not
use that cert to verify the clients.

Add integration tests to validade the certs are in place and can be
used correctly.