slurmd: add singularity-install action
Add an action on charm-slurmd to install singularity using the
resources singularity-deb and singularity-rpm, both also included
in this commit.

The .deb (singularity-ce_3.10.2-focal_amd64.deb) and
.rpm (singularity-ce-3.10.2-1.el7.x86_64.rpm) file should be
downloaded from Singulary Github Releases.

Note: The charm-slurmd/requirements.txt must be updated to use
the version of slurm-ops-manager with the Singulary class
implemented, before applying the action.